@@ -1,11 +1,14 @@
 module github.com/oneclickvirt/memorytest
 
-go 1.22.4
+go 1.23.0
+
+toolchain go1.24.3
 
 require (
 	github.com/oneclickvirt/dd v0.0.1-20250406062523
 	github.com/oneclickvirt/defaultset v0.0.2-20240624082446
-	github.com/oneclickvirt/mbw v0.0.1-20250630131952
+	github.com/oneclickvirt/mbw v0.0.1-20250630140849
+	golang.org/x/sys v0.33.0
 )
 
 require (
